Split geo app's tests into several files.
